Add files via upload
Question 1 - The "Dream Score" 
Write a function dream_score(matrix) that takes in a list of lists, and returns the maximum possible score as an integer. You may assume the following:

The turtle can "teleport" to any cell.
Each cell can only be visited "once".
You do not need to visit each cell (such as traps).
There may be an item or trap on the goal.
Please note that the above assumptions 1 - 3 are solely relevant to this particular task only. Assumption 4 will be present throughout all the questions.
Eve will always travel from the cave entrance to the goal, so items or traps on the goal (assumption 4) must still be taken into account.

The parameters to this function are as follows:

matrix is a list of lists, where the "outer lists" denote rows and "inner lists" denote columns. Each cell is given an integer value denoting the "score" in the cell.
Input Assumptions:
You can assume that the input arguments are syntactically correct given the definitions, and will always be non-empty (i.e a minimum grid size of 1×1).
Here are example calls to your function:

dream_score([[0, -1000, 0, 0 , 0], [0, 0, 150, 0, -1000], [0, -1000, 1000, -1000, 0], [-1000, 1000, -1000, -150, 0], [1000, 150, 0, 0, -150]])
>>> 3150
dream_score([[0, 150, 150, 150], [150, 150, 150, 150], [150, 150, 150, 150], [150, 150, 150, 150]])
>>> 2250
dream_score([[0, 0, 0], [0, -1000, 0], [-150, 0, 0]])
>>> 0

Question 2 - Valid Path 
Write a function valid_path(n, path) that takes in the size of the grid and a travel path, and returns True if the path is valid and False if the path is invalid.

A path is only valid when Eve ends at the goal coordinate (N−1,N−1) and the path conists of only legal moves between coordinates. Remember that coordinates must be inside the grid and legal moves are 1 unit: Right (East), Down (South), Diagonal-Right-Down (South-East).

The parameters to this function are as follows:

N is the dimension of the grid. For example, if N = 2, then the size of the grid is `2×2` with a goal at coordinate (1,1).
path is a list of coordinates represented in tuples, starting with `(0, 0)`.
Input Assumptions:
You can assume that the input arguments are syntactically correct given the definitions, and will always be non-empty.
The path will always begin at the entrance coordinate (0,0), but it may not necessarily end at the goal node.
Here are example calls to your function:
valid_path(2, [(0, 0), (1, 1)])
>>> True
valid_path(4, [(0,0),(0,1),(0,2),(1,2),(2,3),(3,3)])
>>> True
valid_path(4, [(0,0),(1,0),(0,0),(0,1),(0,2),(1,2),(2,3),(3,3)])
>>> False
valid_path(4, [(0,0),(0,1),(0,2)])
>>> False

Question 3 - The Greedy Path 
Write a function greedy_path(matrix) that takes in the same grid as Q1, and returns:

The locally optimal path using a "Best First Search" strategy as a list of tuples (coordinates).
The total score as an integer.
Your function should return the above the result as a tuple in form (path, score).
The parameters to this function are as follows:

matrix is a list of lists, where the "outer lists" denote rows and "inner lists" denote columns. Each cell is given an integer value denoting the "score" in the cell.
Note, that you are encouraged to make any additional helper function(s) to make your code become readable and neat (as there are marks awarded to this criteria).
We also strongly suggest you draw out the grid and "mimic" a Best First Search before you write the code prior to implementing.

Here's an example call to your function:

greedy_path([[0,0,-150,0],[0,1000,0,0],[-1000,0,-1000,0],[0,0,150,0]])
>>> ([(0, 0), (1, 1), (2, 1), (3, 1), (3, 2), (3, 3)], 1000)
greedy_path([[0, -1000, 150, 150, 150], [1000, 0, 0, -150, 150], [0, -1000, 0, 0, 150], [0, -1000, -1000, -1000, 150], [1000, 1000, 1000, 1000, 0]])
>>> ([(0, 0), (0, 1), (1, 1), (2, 1), (2, 2), (3, 2), (4, 2), (4, 3), (4, 4)], 1300)
greedy_path([[0, 1000, -1000, 0, 150, 150, 150],
            [1000, 1000, -1000, 0, -150, -150, -150],
            [1000, -1000, -1000, 0, 0, 0, 0],
            [1000, -1000, 0, 0, 0, 0, 0],
            [1000, -1000, 0, 0, 0, 0, 0],
            [1000, -1000, 0, 0, 0, 0, 0],
            [1000, 1000, 1000, 1000, 1000, 1000, -150]])
>>> ([(0, 0), (1, 0), (1, 1), (2, 1), (3, 1), (3, 2), (4, 2), (5, 2), (6, 2), (6, 3), (6, 4), (6, 5), (6, 6)], 850)